gcc に変更
