+++ .clangd
@@ -1,0 +1,2 @@
+CompileFlags:
+  Remove: -fno-lifetime-dse
