+++ CMakeLists.txt
@@ -23,8 +23,9 @@ include_directories(${PROJECT_BINARY_DIR}/mokey/include)
 
 # llvm::clにてlibc++を使うとSEGVをキメます なんだそれ
 # libstdc++はだいじょぶらしい
-add_compile_options(-stdlib=libc++)
-add_link_options(-stdlib=libc++)
+# pybind11はlibstdc++で死にます フシギダネ
+# add_compile_options(-stdlib=libc++)
+# add_link_options(-stdlib=libc++)
 
 add_subdirectory(mokey)
 add_subdirectory(mevil)
