+++ mevil/CMakeLists.txt
@@ -53,7 +53,7 @@ add_llvm_executable(mevil
         src/main.cpp
         src/runPython.cpp
         src/util/util.cpp
-        src/loadlib/dialect.cpp
+        src/loadlib/load_test.cpp
 )
 
 target_link_libraries(mevil PRIVATE pybind11::embed CLI11::CLI11 dylib)
