+++ mevil/src/loadlib/load_test.h
@@ -5,5 +5,6 @@
 namespace mevil {
 
 void load_dialect_test(std::string path);
+void load_pass_test(std::string path);
 
 }
